refactor: renamed tests folder to test_utils and moved to elasticai/creator
