+++ elasticai/creator/base_modules/linear_test.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 import torch
 from torch.nn.parameter import Parameter
 
-from tests.tensor_test_case import TensorTestCase
+from elasticai.creator.test_utils.tensor_test_case import TensorTestCase
 
 from .linear import Linear, MathOperations
 from .torch_math_operations import TorchMathOperations
