+++ elasticai/creator/base_modules/lstm_test.py
@@ -3,7 +3,7 @@ from typing import Any, Optional, cast
 
 import torch
 
-from tests.tensor_test_case import TensorTestCase
+from elasticai.creator.test_utils.tensor_test_case import TensorTestCase
 
 from .lstm import LSTM
 from .lstm_cell import LSTMCell
