+++ elasticai/creator/nn/binary/_binary_quantization_function_test.py
@@ -4,7 +4,7 @@ from typing import cast
 import torch
 from torch import Tensor, tensor
 
-from tests.tensor_test_case import TensorTestCase
+from elasticai.creator.test_utils.tensor_test_case import TensorTestCase
 
 from ._binary_quantization_function import Binarize
